requirements.txt
dill~=0.3.4
h5py~=3.6.0
jsonpickle~=2.2.0
matplotlib~=3.5.1
meshio~=5.3.4
mpmath~=1.2.1
natsort~=8.1.0
numpy~=1.21.2
Ofpp~=0.11
pandas~=1.4.1
plotly~=5.1.0
python_box~=5.4.1
pyvista~=0.36.1
PyYAML~=6.0
scikit_learn~=1.1.2
scipy~=1.8.0
seaborn~=0.11.2
setuptools~=60.9.3
Shapely~=1.8.4
streamlit~=1.12.0
tabulate~=0.8.10
tqdm~=4.62.3
translate~=3.6.1
trimesh~=3.14.0
vtk~=9.1.0
